修正了README api文档以comId及条件判断

+++ src/main/java/org/example/sastwoc/controller/CaptainController.java
@@ -1,5 +1,6 @@
 package org.example.sastwoc.controller;
 
+import io.swagger.v3.oas.models.security.SecurityScheme;
 import org.example.sastwoc.DTO.DeleteRequest;
 import org.example.sastwoc.DTO.MemberInformation;
 import org.example.sastwoc.entity.Member;
@@ -12,6 +13,7 @@ import org.example.sastwoc.service.AdminService;
 import org.example.sastwoc.service.CaptainService;
 import org.example.sastwoc.utils.JwtUtils;
 import org.example.sastwoc.vo.Result;
+import org.example.sastwoc.vo.Status;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
@@ -37,13 +39,29 @@ public class CaptainController {
     //添加成员
     @PostMapping("/team/{teamId}/member")
     public Result teamMemberAdd(@PathVariable Integer teamId,@RequestBody Member member){
+        String authorization=httpServletRequest.getHeader("Authorization");
+        String token = authorization.replace("Bearer ", "");
+        String userCode = JwtUtils.resolveJwt(token).getUser_code();
+        Integer myTeamId = personMapper.getTeamIdByUserCode(userCode);
+        if(myTeamId.intValue()!=teamId.intValue()){
+            throw new SystemException(Status.USER_ACCESS_NOT_ALLOWED);
+        }
          captainService.addMember(member,teamId);
          return Result.success();
     }
 
     //删除成员
     @DeleteMapping("/team/{teamId}/member")
-    public Result teamMemberDelete(@PathVariable String teamId, @RequestBody DeleteRequest request){
+    public Result teamMemberDelete(@PathVariable Integer teamId, @RequestBody DeleteRequest request){
+        String authorization=httpServletRequest.getHeader("Authorization");
+        String token = authorization.replace("Bearer ", "");
+        String userCode = JwtUtils.resolveJwt(token).getUser_code();
+        Integer myTeamId = personMapper.getTeamIdByUserCode(userCode);
+        if(myTeamId.intValue()!=teamId.intValue()){
+            throw new SystemException(Status.USER_ACCESS_NOT_ALLOWED);
+        }
+
+
         int id = request.getId();
             captainService.deleteMember(id);
             return Result.success();
@@ -51,7 +69,18 @@ public class CaptainController {
 
     //修改成员信息
     @PatchMapping("/team/{teamId}/member")
-    public Result teamMemberUpdate(@PathVariable String teamId,@RequestBody Member member){
+    public Result teamMemberUpdate(@PathVariable Integer teamId, @RequestBody Member member){
+        String authorization=httpServletRequest.getHeader("Authorization");
+        String token = authorization.replace("Bearer ", "");
+        String userCode = JwtUtils.resolveJwt(token).getUser_code();
+        Integer myTeamId = personMapper.getTeamIdByUserCode(userCode);
+        if(myTeamId.intValue()!=teamId.intValue()){
+            throw new SystemException(Status.USER_ACCESS_NOT_ALLOWED);
+        }
+
+
+        member.setTeamId(teamId);
+        member.setIsCaptain(0);
         captainService.editMember(member);
         return Result.success();
     }
@@ -59,6 +88,15 @@ public class CaptainController {
     //获取成员列表
     @GetMapping("/team/{teamId}/member")
     public Result teamMemberGet(@PathVariable Integer teamId){
+        String authorization=httpServletRequest.getHeader("Authorization");
+        String token = authorization.replace("Bearer ", "");
+        String userCode = JwtUtils.resolveJwt(token).getUser_code();
+        Integer myTeamId = personMapper.getTeamIdByUserCode(userCode);
+        if(myTeamId.intValue()!=teamId.intValue()){
+            throw new SystemException(Status.USER_ACCESS_NOT_ALLOWED);
+        }
+
+
         List<MemberInformation> memberList=new ArrayList<>();
         memberList=captainService.getMemberList(teamId);
         return Result.success(memberList);
@@ -67,7 +105,13 @@ public class CaptainController {
     //获取队伍信息
     @GetMapping("/team")
     public Result teamGet(){
-        String userCode = (JwtUtils.resolveJwt(httpServletRequest.getHeader("token"))).getUser_code();
+        //String userCode = (JwtUtils.resolveJwt(httpServletRequest.getHeader("token"))).getUser_code();
+
+        String authorization=httpServletRequest.getHeader("Authorization");
+        String token = authorization.replace("Bearer ", "");
+        String userCode = JwtUtils.resolveJwt(token).getUser_code();
+
+
         Integer teamId=personMapper.getTeamIdByUserCode(userCode);
         TeamModel teamModel=new TeamModel();
         teamModel=captainService.getTeamInformation(teamId);
